Add the ability to erase a drive without first setting a password.
When using the default password (MSID), the password should not be
hashed when establishing the session with the drive.  To indicate
that sedutil-cli should use the default password (and not hash it),
pass in a NULL string for the password "".

This allows wiping a drive that has not been password protected.

Common/DtaDevEnterprise.cpp:
	In eraseLockingRange(), if the user passes in a NULL string
	for the password, fetch the MSID and setup the session to
	not hash the password sent.